Begin implementing new API
Moving API to AccuWeather for better accuracy
In progress
Todo:
- fix response parsing
- reduce number of calls per day
- change the way rain conditions are handled (not using history and current)

+++ include/helpers.h
@@ -7,6 +7,6 @@
 #define HIGH_RAIN 100
 
 double getPrecipitation(char *response);
-int updateLeds(double current[6], double history[6], int handle, int leds[6]);
+int updateLeds(int current[6], double history[6], int handle, int leds[6]);
 
 #endif // HELPERS_H
